LayoutTests/imported/w3c/web-platform-tests/svg/interact/scripted/text-hittest-002-expected.txt

MMMMMM

FAIL elementFromPoint(...) on <svg:text> with fill=none but fill=black descendants assert_equals: element @ (175, 25) expected Element node <tspan fill="black">MM</tspan> but got Element node <svg width="400" height="400">
  <text x="50" y="50" font...
